requirements.txt
torch>=2.1
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
requests>=2.31
